Separate symbol display, add symbols animation
